+++ lib/subscription/constants.ts
@@ -1,0 +1,12 @@
+// UI 표시용 옵션들
+
+// 결제 방법 옵션
+export const PAYMENT_OPTIONS = [
+  { value: 'CARD', label: '신용카드' },
+  { value: 'VIRTUAL_ACCOUNT', label: '가상계좌' },
+  { value: 'TRANSFER', label: '실시간 계좌이체' },
+];
+
+// 배송비 기준
+export const FREE_SHIPPING_THRESHOLD = 50000;
+export const SHIPPING_FEE = 2500;
