+++ package.json
@@ -11,10 +11,13 @@
     "format:check": "prettier . --check"
   },
   "dependencies": {
+    "@portone/browser-sdk": "^0.1.3",
     "next": "16.1.1",
     "pretendard": "^1.3.9",
     "react": "19.2.3",
-    "react-dom": "19.2.3"
+    "react-dom": "19.2.3",
+    "react-spinners": "^0.17.0",
+    "zustand": "^5.0.10"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4",
